First working commit after SDL update and platform change.
A minimal window with the full sprite sheet loaded is rendered
correctly. No other functionality is present yet. The main goal here was
to update to SDL 2.0 and switch to a Linux platform. Man, it's nice to
not have a DLL folder!!

Much of the structure I had in the last version is still present though
isn't being used at the moment. It's still going to be awhile before I
have any idea what I'm doing.