Update masks.png.
No more blurred lines! Used a different algorithm to stretch it so that
it's pixelated and not blurred. This cut the file size down to nothing
and made it look better for this purpose.

Trying to keep image updates in their own commit so that, if necessary,
I can go back and delete the commits later.