Update set_tiles.
Modified set_tiles so that it uses a new method of loading tiles from a
file. This is based on what my level editor spits out. This made it so I
could get rid of all of the hardcoded source values.

Known bug: There seems to be an off-by-one error somewhere in the tile
loading. Everything is 32 pixels to the left of where it should be, and
some tiles are getting sourced wrong too.