Implement WALKR animation bound to key.
Cleaned up quite a bit of the mess, but still reconsidering the still
very messy data structures.

One of the greater challenges I'm facing now is the framerate and how it
affects everything. I added a FRAMETIME constant that represents the
number of timer loops per frame, but this isn't quite enough as this
also affects non-animation values like velocity.