Detect a lack of keyboard input.
This is useful to determine whether any non-idle animation is being
played, that way the idle animation doesn't overwrite another. I would
much prefer a better way of handling this so that it doesn't involve
assigning the idle variable constantly. A better way would be to detect
if ANY key is being pressed, but I don't know how to do this.