Add Controller class.
The purpose of the controller class is to allow global messaging between
the game loop and the game logic. I'd like to detect keyup events so
that I can tell when a key has been released. This should get some of
the annoying controller logic outside of the game logic and hopefully
make things a little cleaner, at the cost of another global variable.

I also deleted unused variables in globals.