Add controller limitations to shooting.
Specifically, don't allow holding the spacebar to keep firing or to keep
using the animation. Also, enforce the animation lasting 4 frames
instead of 1. This will probably need tweaking so that it doesn't lock
the player up, for example, during a shoot() while jumping and moving
left/right.

There's still lots of cleanup to be done.