Add jump shooting left/right animations.
The animation got a little tricky because of the order of operations.
I set it up so that one animation can overwrite another based on
precedence, for example, jump shooting gets priority over falling, but
only if we're shooting.

I noticed a movement bug while jump shooting. In the actual game, I
think there might be xAccel while you're in the air, but not on the
ground. This makes it so you can jump right, shoot, and keep moving
right without holding right. Currently xVel gets set to 0 if I shoot.