Fix mid-air movement bug (mostly).
This got a little messy, and I need to rewrite some things to make it
cleaner. I started using xAccel so that I can move in mid-air with air
resistance. It also allows me to shoot in mid-air but keep moving
afterward. Compared to the real game, the initial push from a movement
key in mid-air (xVel = 5) is a bit strong, and the drag is a bit strong
too. I need a way to balance these out, but at the moment that's a
conflicting idea because I've done it linearly.