Experiment with changing animState to int rather than enum.
This lets me get rid of all the annoying if statements to figure out
whether to do the LEFT or RIGHT variant of the srcClip. The downside is
now the names aren't friendly. A way around this would be to \#define
some constants for this.

Noticed a bug: Now I can shoot while walking on the ground. Shouldn't be
able to.