Fix bug where you could double jump.
This was possible because bottom collision is only checked (and thus
isOnGround is only set) if yVel > 0, and it isn't when you jump. The fix
is naive though, because it assumes nothing happens to stop the jump
before it actually happens. It would be nice to have a better way of
determining whether we're on the ground.