Put BlasterShot on screen with animation.
Nothing useful yet, just a test of getting the shot on the screen.

Going to work on refactoring a bit. Going to put the animation data in a
struct, rather than floating around outside of the class. The struct
will then be put in the class. The point of this is to avoid all of that
data from being entered explicitly in both the .h and .cpp files.