Leave animation out of Sprite for now.
I'd like to put animate() into Sprite because it should be able to apply
to any Sprite very easily, but the anims array is giving me trouble
since it's a statically allocated array. Having it static makes sense
because it loads predictable, static data, but not if it has to provide
animations for multiple classes.

In the future maybe I can reconsider this data structure OR split
animate() into two functions and have one of them exist in each derived
class.