Refactor player-keyboard input logic.
All keyboard input that controls the player is now inside
processKeyboard(). This makes it easy to only process keyboard actions
if no "blocking" action (shooting, hanging from a ledge, etc.) is
already occurring. Before this commit, you could shoot and then jump
while still in the shoot animation. Now you can't.

Not only does this fix the bug, but it's also much simpler and easier to
read. It also provides for expansion in the future.