Allow jump while on pogo.
Not really "working" yet. Still need to adjust power of each jump and
there's an odd bug occurring where the idle pogo will occasionally not
reach its full height even with nothing in the way.

A big thing on the todo list is to use floats in hitboxes and rendering
for more precision. This precision is absolutely necessary for
jumping/falling to work smoothly because the integer values are too
small to provide enough precision.