Tweak pogo jump to match new physics.
It's pretty close to what I want but not quite. The idle pogo is quite a
bit too fast because of the increased gravity. I need a way around this
but modifying gravity just while on a pogo stick is not my first choice.

I think the idle pogo height and the max height are just about what I
want.