Improve x-direction falling velocities.
Haven't figured this one out yet. I need drag to be pretty strong but
also for Keen to build quite a bit of momentum if an arrow key is held
one direction for awhile. This is most noticeable in-game if you jump to
the right for awhile then start holding left. It takes a long time to
make any progress in the other direction. Mine doesn't have this quality
and it'll take some more experimenting to figure out.

One of the large problems is with storing xVelRem for drag. What
*should* happen is, if moving right, xVelRem should be positive until
you let go of the arrow key, then it should zero out, then start getting
negative. The problem is it stays positive. I've temporarily fixed this
by just making it zero once drag is applied, but this isn't quite right.
I think I need some sort of "this is the first loop without a keypress"
flag so that I can zero the remainder just once, but that's kind of
messy.