Experiment with pole detection.
I'm trying out a new approach to collision. Instead of just running
"check collision", depending on what you want to know, you get tiles to
check in some direction. The downside to this is that it has to create
and return a vector of tile pointers then loop through it later, which
is more work than really necessary.

At this point, this new system is only used for poles. If it works I can
bring it to all other tiles too. I think my main concern is the
performance overhead which probably isn't a big deal, but seems
unnecessarily wasteful.