Correct first tile checked for in any direction.
The fact that you can be "inside" pole tiles but still not colliding
with them made this one obvious. The downside is that the way it WAS
provided an advantage: it made it easier to detect spawning problems
(such as shooting past a tile). Now I have this bug again where you can
shoot past tiles so I'm faced with two choices:
1) Rollback this commit and adjust the pole collision accordingly
2) Find another way around the spawning bug