Support starting pole climb in DOWN direction.
Accomplished this with a quick and hopefully temporary messy solution. I
added a parameter to GetTileCollidingWithBottom to check "IsTouching"
instead of "IsColliding". Ideally these would be two distinct functions
but I didn't want all of the code duplication involved. It's probably
for the best later though.

I only implemented this for bottom collision until it's required by more
later.