Add unit collision detection to BlasterShot.
I chose to detect collision from the BlasterShot's perspective (rather
than the enemy's perspective) because it'll only need to be in one class
instead of several, and because no dynamic casts are currently
involved...it can instead use the Sprite pointer directly.

There are a couple problems at the moment...
1) In some frames the player can hit himself with the shot because of
the shot spawn
2) The shot is always colliding with itself and this is always being
checked

This leaves me with a few obvious options:
1) Do collision from the enemy's perspective
2) Add exceptions to the shot and the player, somehow preventing either
of them from being "shot"
3) Split up the sprite batch into multiple batches. Currently I'm
thinking Player, BlasterShots, and Enemies. I think this would be fine
as long as there weren't any more than 3 separate groups.