Add climbing cooldown timer.
This fixes the bug where Ampton would get stuck on a pole on the ground.
Sadly there's still a small bug remaining where Ampton walks past a pole
on the ground...he doesn't realize it's the end of the pole until he's
already tried to climb it. Kind of hard to get around that.