Get platform movement with player mostly working.
There are a couple problems I haven't solved yet.
When the platform moves up, it causes collision with the player.
When the platform moves down, it moves away from colliding with the
player.
Each of these situations creates a unique problem and it seems like
solving one breaks the other.