Change which tile is the edge tile on the left.
I made it so that the tile with only a little art in it is just for
show. It has no properties. The tile directly to its right is now the
edge tile. The confusing part of this is that that tile looks exactly
the same as the "normal" tile and so I need to make a big distinction
between the two when designing levels.