Implement temporary fix for slope collision.
This along with the last commit should be considered a temporary hacky
solution to a larger collision problem. I've given a couple tiles slope
properties so that they will count bottom collision when entering from
the sides. This is important for tiles at the end of a slope and
otherwise results in Keen getting stuck when proceeding from a slope to
flat ground.