Make Ampton change direction at edge of pole.
The remaining bug is that Ampton never goes back to patrolling on the
pole he's currently being tested on because not enough time elapses
between his direction change and the patrol check. To fix this, I need
to find a way to prevent premature direction changes other than by using
an elapsed time.