Add EnemyLaserManager and automate enemy laser shots.
This works but there are a couple hacks I need to fix.

1. The enemy lasers are added based on the source image on the sprite
sheet rather than actual level data. It would be a pain to add this to
the level data but that may be better than hardcoding an x,y coordinate
indicating that a stationary gun is here.

2. Currently Keen's BlasterShot frames are being used because I don't
have a good way of choosing which texture to use based on what type of
laser it is. My plan is to try to store the texture pointer in each
class rather than passing it in to each draw() call. This should result
in cleaner code, too.