Use correct sprite sheet for EnemyLaser.
A couple notes:

1. Ideally there would only be one sprite sheet. This is generally
better but I happened to start writing this with two because I
downloaded them online and never combined them. Thankfully this doesn't
present a performance problem but if I had to do it over again, there
would only be one sprite sheet.

2. A long-term goal is to make all draw() methods not take a Texture*
parameter but instead have the Texture* stored in the class. This
would be pointless if I only had one sprite sheet but that's not likely
to happen.