Generate make dependencies automatically.
Previously whenever I changed a .h file I had to make clean and make
again. This got to be kind of time consuming as my full build sits
around 6 seconds (I'm using a VM, but that still seems a little slow
given the small size of this program).

What I still don't understand is that the .DEFAULT_GOAL isn't what I
expect. It gets set to the first object file (ampton.o currently) but
only checks that file, and not the rest of the object files. I expect
.DEFAULT_GOAL to be 'all', but it isn't.

I really don't understand make.